Update check-files and error annotations
.. of tests explicitly checking implicits resolving to most specific
