+++ tests/run/enrich-gentraversable.check
@@ -4,5 +4,5 @@ List(2, 4)
 HW
 ArraySeq(72, 108, 108, 32, 114, 108, 100)
 Map(bar -> 2)
-TreeMap(bar -> 2)
+Map(bar -> 2)
 Map(bar -> 2)
