+++ tests/run/implicits_poly.check
@@ -1,1 +1,1 @@
-barFoo
+fooFoo
